+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Handle dwebp re-coding errors
+
 ## [3.10.0] - 17.06.2024
 
 ### Added